pystartowo/requirements.txt
Django~=4.1
django-environ~=0.9.0
# 
# NOTE: if you have trouble installing psycopg2 while developing,
# consider changing your requirement to use `psycopg2-binary`
# See: https://www.psycopg.org/docs/install.html#psycopg-vs-psycopg-binary
psycopg2-binary~=2.9.3
# 
